writing v1 of docs
